Added unit test for lookup service
